refactor(analysis): extract derived distribution owner
